Add gi-assets-data with filenames for assets
