+++ libs/gi/assets-data/README.md
@@ -1,0 +1,3 @@
+# gi-assets-data
+
+This library was generated with [Nx](https://nx.dev).
