+++ libs/gi/assets-data/project.json
@@ -1,0 +1,9 @@
+{
+  "name": "gi-assets-data",
+  "$schema": "../../../node_modules/nx/schemas/project-schema.json",
+  "sourceRoot": "libs/gi/assets-data/src",
+  "projectType": "library",
+  "tags": [],
+  "// targets": "to see all targets run: nx show project gi-assets-data --web",
+  "targets": {}
+}
