+++ tsconfig.base.json
@@ -53,6 +53,7 @@
       "@genshin-optimizer/common/util": ["libs/common/util/src/index.ts"],
       "@genshin-optimizer/gi/art-scanner": ["libs/gi/art-scanner/src/index.ts"],
       "@genshin-optimizer/gi/assets": ["libs/gi/assets/src/index.ts"],
+      "@genshin-optimizer/gi/assets-data": ["libs/gi/assets-data/src/index.ts"],
       "@genshin-optimizer/gi/char-cards": ["libs/gi/char-cards/src/index.ts"],
       "@genshin-optimizer/gi/consts": ["libs/gi/consts/src/index.ts"],
       "@genshin-optimizer/gi/db": ["libs/gi/db/src/index.ts"],
